Minor changes to Gibbs sampler C++ code to allow sampling in steps of 100 sweeps instead of 1000 sweeps before saving intermediate results

+++ maxent/maxEnt_gibbs_pair_C.m
@@ -53,7 +53,7 @@ switch machine
                                    double(xc), pairs-1, m, fm-1, h, J, L);
     case 'laptop'
 [xSampled,xc] = pwGibbsMaxEnt_malloc(int32(nSamples),int32(burnIn),int32(d),...
-                                   xc, pairs-1, m, fm-1, h, J, L);
+                                   double(xc), pairs-1, m, fm-1, h, J, L);
  E = [];       
     otherwise
 [xSampled,xc] = pwGibbsMaxEnt_malloc(int32(nSamples),int32(burnIn),int32(d),...
